Moved some data classes to there own folder
