+++ TowerDefence_UnityProject/Packages/TowerDefence/World/Grid/Data/IGridNode.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace TowerDefence.World.Grid
+namespace TowerDefence.World.Grid.Data
 {
     internal interface IGridNode
     {
